+++ include/UDRefl/ReflMngr.h
@@ -244,6 +244,7 @@ namespace Ubpa::UDRefl {
 		//
 
 		// parameter <- argument
+		// - ObjectView
 		// - same
 		// - reference
 		// > - 0 (invalid), 1 (convertible), 2 (constructible)